+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - CoffeeNet Starter Parent to version `0.37.0`
+- GroupId to 'rocks.coffeenet'
 
 ## [0.5.0]
 ### Changed